requirements.txt
pytest
psycopg2-binary
records

